Reorganized namespacing.
Added a Request namespace for all AbstractRequest implementations.
