+++ clients/csharp/src/Kafka/Kafka.Client/Request/MultiFetchRequest.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using Kafka.Client.Util;
 
-namespace Kafka.Client
+namespace Kafka.Client.Request
 {
     /// <summary>
     /// Constructs a multi-consumer request to send to Kafka.
