+++ clients/csharp/src/Kafka/Tests/Kafka.Client.Tests/Kafka.Client.Tests.csproj
@@ -44,14 +44,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FetchRequestTests.cs" />
-    <Compile Include="MultiFetchRequestTests.cs" />
     <Compile Include="KafkaIntegrationTest.cs" />
     <Compile Include="MessageTests.cs" />
-    <Compile Include="MultiProducerRequestTests.cs" />
-    <Compile Include="OffsetRequestTests.cs" />
-    <Compile Include="ProducerRequestTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Request\FetchRequestTests.cs" />
+    <Compile Include="Request\MultiFetchRequestTests.cs" />
+    <Compile Include="Request\MultiProducerRequestTests.cs" />
+    <Compile Include="Request\OffsetRequestTests.cs" />
+    <Compile Include="Request\ProducerRequestTests.cs" />
     <Compile Include="Util\BitWorksTests.cs" />
   </ItemGroup>
   <ItemGroup>
