+++ clients/csharp/src/Kafka/Tests/Kafka.Client.Tests/KafkaIntegrationTest.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Threading;
+using Kafka.Client.Request;
 using NUnit.Framework;
 
 namespace Kafka.Client.Tests
