+++ clients/csharp/src/Kafka/Tests/Kafka.Client.Tests/Request/OffsetRequestTests.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Linq;
 using System.Text;
+using Kafka.Client.Request;
 using Kafka.Client.Util;
 using NUnit.Framework;
 
-namespace Kafka.Client.Tests
+namespace Kafka.Client.Request.Tests
 {
     /// <summary>
     /// Tests the <see cref="OffsetRequest"/> class.
