+++ clients/csharp/src/Kafka/Tests/Kafka.Client.Tests/Request/ProducerRequestTests.cs
@@ -2,10 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Kafka.Client.Request;
 using Kafka.Client.Util;
 using NUnit.Framework;
 
-namespace Kafka.Client.Tests
+namespace Kafka.Client.Request.Tests
 {
     /// <summary>
     /// Tests for the <see cref="ProducerRequest"/> class.
